+++ code/src/cljs/sixsq/nuvla/ui/utils/general.cljs
@@ -371,3 +371,7 @@
     {\- "\\-", \[ "\\[", \] "\\]", \{ "\\{", \} "\\}",
      \( "\\(", \) "\\)", \* "\\*", \+ "\\+", \? "\\?",
      \. "\\.", \\ "\\\\", \^ "\\^", \$ "\\$", \| "\\|"}))
+
+(defn envsubst-str
+  [envsubst s]
+  (str/replace s (->> envsubst keys (map regex-escape) (str/join "|") re-pattern) envsubst))